api/requirements.txt
# API server dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sse-starlette>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
supabase>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-multipart>=0.0.9
